refactor: isolate australia universe dispatch
